Communication between Currency L0 and Global L0
