--- modules/currency-l0/src/main/scala/org/tessellation/currency/config/types.scala
@@ -1,15 +1,0 @@
-package org.tessellation.currency.config
-
-import org.tessellation.sdk.config.AppEnvironment
-import org.tessellation.sdk.config.types._
-
-object types {
-  case class AppConfig(
-    environment: AppEnvironment,
-    http: HttpConfig,
-    gossip: GossipConfig,
-    healthCheck: HealthCheckConfig,
-    snapshot: SnapshotConfig,
-    collateral: CollateralConfig
-  )
-}
