--- modules/currency-l0/src/main/scala/org/tessellation/currency/domain/cell/AlgebraCommand.scala
@@ -1,12 +1,0 @@
-package org.tessellation.currency.domain.cell
-
-import org.tessellation.currency.schema.currency.CurrencyBlock
-import org.tessellation.kernel.Ω
-import org.tessellation.security.signature.Signed
-
-sealed trait AlgebraCommand extends Ω
-
-object AlgebraCommand {
-  case class EnqueueL1BlockData(data: Signed[CurrencyBlock]) extends AlgebraCommand
-  case object NoAction extends AlgebraCommand
-}
