--- modules/currency-l0/src/main/scala/org/tessellation/currency/domain/cell/CoalgebraCommand.scala
@@ -1,10 +1,0 @@
-package org.tessellation.currency.domain.cell
-
-import org.tessellation.currency.schema.currency.CurrencyBlock
-import org.tessellation.security.signature.Signed
-
-sealed trait CoalgebraCommand
-
-object CoalgebraCommand {
-  case class ProcessL1Block(data: Signed[CurrencyBlock]) extends CoalgebraCommand
-}
