--- modules/currency-l0/src/main/scala/org/tessellation/currency/http/P2PClient.scala
@@ -1,38 +1,0 @@
-package org.tessellation.currency.http
-
-import cats.effect.Async
-
-import org.tessellation.kryo.KryoSerializer
-import org.tessellation.sdk.domain.cluster.services.Session
-import org.tessellation.sdk.http.p2p.SdkP2PClient
-import org.tessellation.sdk.http.p2p.clients._
-import org.tessellation.sdk.infrastructure.gossip.p2p.GossipClient
-import org.tessellation.security.SecurityProvider
-
-import org.http4s.client.Client
-
-object P2PClient {
-
-  def make[F[_]: Async: SecurityProvider: KryoSerializer](
-    sdkP2PClient: SdkP2PClient[F],
-    client: Client[F],
-    session: Session[F]
-  ): P2PClient[F] =
-    new P2PClient[F](
-      sdkP2PClient.sign,
-      sdkP2PClient.cluster,
-      sdkP2PClient.gossip,
-      sdkP2PClient.node,
-      sdkP2PClient.trust,
-      GlobalSnapshotClient.make[F](client, session)
-    ) {}
-}
-
-sealed abstract class P2PClient[F[_]] private (
-  val sign: SignClient[F],
-  val cluster: ClusterClient[F],
-  val gossip: GossipClient[F],
-  val node: NodeClient[F],
-  val trust: TrustClient[F],
-  val globalSnapshot: GlobalSnapshotClient[F]
-)
