+++ modules/currency-l0/src/main/scala/org/tessellation/currency/l0/cell/CoalgebraCommand.scala
@@ -1,0 +1,10 @@
+package org.tessellation.currency.l0.cell
+
+import org.tessellation.currency.schema.currency.CurrencyBlock
+import org.tessellation.security.signature.Signed
+
+sealed trait CoalgebraCommand
+
+object CoalgebraCommand {
+  case class ProcessL1Block(data: Signed[CurrencyBlock]) extends CoalgebraCommand
+}
