+++ modules/currency-l0/src/main/scala/org/tessellation/currency/l0/modules/Queues.scala
@@ -1,0 +1,28 @@
+package org.tessellation.currency.l0.modules
+
+import cats.effect.Concurrent
+import cats.effect.std.Queue
+import cats.syntax.functor._
+
+import org.tessellation.currency.schema.currency.CurrencyBlock
+import org.tessellation.schema.gossip.RumorRaw
+import org.tessellation.sdk.modules.SdkQueues
+import org.tessellation.security.Hashed
+import org.tessellation.security.signature.Signed
+
+object Queues {
+
+  def make[F[_]: Concurrent](sdkQueues: SdkQueues[F]): F[Queues[F]] =
+    for {
+      l1OutputQueue <- Queue.unbounded[F, Signed[CurrencyBlock]]
+    } yield
+      new Queues[F] {
+        val rumor: Queue[F, Hashed[RumorRaw]] = sdkQueues.rumor
+        val l1Output: Queue[F, Signed[CurrencyBlock]] = l1OutputQueue
+      }
+}
+
+sealed abstract class Queues[F[_]] private {
+  val rumor: Queue[F, Hashed[RumorRaw]]
+  val l1Output: Queue[F, Signed[CurrencyBlock]]
+}
