--- modules/currency-l0/src/main/scala/org/tessellation/currency/modules/Programs.scala
@@ -1,30 +1,0 @@
-package org.tessellation.currency.modules
-
-import cats.effect.Async
-
-import org.tessellation.sdk.domain.cluster.programs.{Joining, PeerDiscovery}
-import org.tessellation.sdk.infrastructure.snapshot.programs.Download
-import org.tessellation.sdk.modules.SdkPrograms
-
-object Programs {
-
-  def make[F[_]: Async](
-    sdkPrograms: SdkPrograms[F],
-    storages: Storages[F],
-    services: Services[F]
-  ): Programs[F] = {
-    val download = Download
-      .make(
-        storages.node,
-        services.consensus
-      )
-
-    new Programs[F](sdkPrograms.peerDiscovery, sdkPrograms.joining, download) {}
-  }
-}
-
-sealed abstract class Programs[F[_]] private (
-  val peerDiscovery: PeerDiscovery[F],
-  val joining: Joining[F],
-  val download: Download[F]
-)
