+++ modules/sdk/src/main/scala/org/tessellation/sdk/domain/block/processing/BlockAcceptanceResult.scala
@@ -3,12 +3,12 @@ package org.tessellation.sdk.domain.block.processing
 import org.tessellation.schema.Block
 import org.tessellation.security.signature.Signed
 
-import derevo.cats.{eqv, show}
+import derevo.cats.show
 import derevo.derive
 import eu.timepit.refined.cats._
 import eu.timepit.refined.types.numeric.NonNegLong
 
-@derive(eqv, show)
+@derive(show)
 case class BlockAcceptanceResult[B <: Block[_]](
   contextUpdate: BlockAcceptanceContextUpdate,
   accepted: List[(Signed[B], NonNegLong)],
